backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg2-asyncpg
alembic
PyJWT
passlib[bcrypt]
pydantic-settings
cryptography
python-dateutil
